Make llvm_ops.cpp leaner by removing most of the matrix functions that are either very rarely called, or that would not benefit much from being inlineable.
