+++ src/liboslexec/CMakeLists.txt
@@ -3,8 +3,7 @@ SET ( liboslexec_srcs
           dictionary.cpp
           context.cpp instance.cpp 
           loadshader.cpp master.cpp 
-          opcolor.cpp 
-          opmessage.cpp
+          opcolor.cpp opmatrix.cpp opmessage.cpp
           opnoise.cpp simplexnoise.cpp gabornoise.cpp
           opspline.cpp opstring.cpp optexture.cpp
           oslexec.cpp osoreader.cpp
