+++ src/liboslexec/builtindecl.h
@@ -200,6 +200,8 @@ DECL (osl_prepend_color_from, "xXXs")
 DECL (osl_prepend_matrix_from, "iXXs")
 DECL (osl_get_matrix, "iXXs")
 DECL (osl_get_inverse_matrix, "iXXs")
+DECL (osl_transform_triple, "iXXiXiXXi")
+DECL (osl_transform_triple_nonlinear, "iXXiXiXXi")
 
 DECL (osl_dict_find_iis, "iXiX")
 DECL (osl_dict_find_iss, "iXXX")
@@ -299,20 +301,6 @@ DECL (osl_transformv_dvmdv, "xXXX")
 DECL (osl_transformn_vmv, "xXXX")
 DECL (osl_transformn_dvmdv, "xXXX")
 
-DECL (osl_transform_triple, "iXXiXiXXi")
-DECL (osl_transform_triple_nonlinear, "iXXiXiXXi")
-
-DECL (osl_mul_mm, "xXXX")
-DECL (osl_mul_mf, "xXXf")
-DECL (osl_mul_m_ff, "xXff")
-DECL (osl_div_mm, "xXXX")
-DECL (osl_div_mf, "xXXf")
-DECL (osl_div_fm, "xXfX")
-DECL (osl_div_m_ff, "xXff")
-DECL (osl_get_from_to_matrix, "iXXss")
-DECL (osl_transpose_mm, "xXX")
-DECL (osl_determinant_fm, "fX")
-
 DECL (osl_dot_fvv, "fXX")
 DECL (osl_dot_dfdvdv, "xXXX")
 DECL (osl_dot_dfdvv, "xXXX")
@@ -330,6 +318,17 @@ DECL (osl_distance_dfvdv, "xXXX")
 DECL (osl_normalize_vv, "xXX")
 DECL (osl_normalize_dvdv, "xXX")
 #endif
+
+DECL (osl_mul_mm, "xXXX")
+DECL (osl_mul_mf, "xXXf")
+DECL (osl_mul_m_ff, "xXff")
+DECL (osl_div_mm, "xXXX")
+DECL (osl_div_mf, "xXXf")
+DECL (osl_div_fm, "xXfX")
+DECL (osl_div_m_ff, "xXff")
+DECL (osl_get_from_to_matrix, "iXXss")
+DECL (osl_transpose_mm, "xXX")
+DECL (osl_determinant_fm, "fX")
 
 DECL (osl_concat_sss, "sss")
 DECL (osl_strlen_is, "is")
