+++ src/liboslexec/llvm_ops.cpp
@@ -94,7 +94,6 @@ examples), as you are just coding in C++, but there are some rules:
 #include <iostream>
 
 #include "OSL/oslconfig.h"
-#include "OSL/rendererservices.h"
 #include "OSL/shaderglobals.h"
 #include "OSL/dual.h"
 #include "OSL/dual_vec.h"
@@ -643,201 +642,6 @@ OSL_SHADEOP void osl_transformn_dvmdv(void *result, void* M_, void* v_)
    const Dual2<Vec3> &v = DVEC(v_);
    const Matrix44    &M = MAT(M_);
    multDirMatrix (M.inverse().transpose(), v, DVEC(result));
-}
-
-
-
-// Matrix ops
-
-OSL_SHADEOP void
-osl_mul_mm (void *r, void *a, void *b)
-{
-    MAT(r) = MAT(a) * MAT(b);
-}
-
-OSL_SHADEOP void
-osl_mul_mf (void *r, void *a, float b)
-{
-    MAT(r) = MAT(a) * b;
-}
-
-OSL_SHADEOP void
-osl_mul_m_ff (void *r, float a, float b)
-{
-    float f = a * b;
-    MAT(r) = Matrix44 (f,0,0,0, 0,f,0,0, 0,0,f,0, 0,0,0,f);
-}
-
-OSL_SHADEOP void
-osl_div_mm (void *r, void *a, void *b)
-{
-    MAT(r) = MAT(a) * MAT(b).inverse();
-}
-
-OSL_SHADEOP void
-osl_div_mf (void *r, void *a, float b)
-{
-    MAT(r) = MAT(a) * (1.0f/b);
-}
-
-OSL_SHADEOP void
-osl_div_fm (void *r, float a, void *b)
-{
-    MAT(r) = a * MAT(b).inverse();
-}
-
-OSL_SHADEOP void
-osl_div_m_ff (void *r, float a, float b)
-{
-    float f = (b == 0) ? 0.0f : (a / b);
-    MAT(r) = Matrix44 (f,0,0,0, 0,f,0,0, 0,0,f,0, 0,0,0,f);
-}
-
-OSL_SHADEOP int osl_get_matrix (void *sg, void *r, const char *from);
-OSL_SHADEOP int osl_get_inverse_matrix (void *sg, void *r, const char *to);
-OSL_SHADEOP int osl_prepend_matrix_from (void *sg, void *r, const char *from);
-
-
-OSL_SHADEOP int
-osl_get_from_to_matrix (void *sg, void *r, const char *from, const char *to)
-{
-    Matrix44 Mfrom, Mto;
-    int ok = osl_get_matrix ((ShaderGlobals *)sg, &Mfrom, from);
-    ok &= osl_get_inverse_matrix ((ShaderGlobals *)sg, &Mto, to);
-    MAT(r) = Mfrom * Mto;
-    return ok;
-}
-
-
-
-OSL_SHADEOP int
-osl_transform_triple (void *sg_, void *Pin, int Pin_derivs,
-                      void *Pout, int Pout_derivs,
-                      void *from, void *to, int vectype)
-{
-    static ustring u_common ("common");
-    ShaderGlobals *sg = (ShaderGlobals *)sg_;
-    Matrix44 M;
-    int ok;
-    Pin_derivs &= Pout_derivs;   // ignore derivs if output doesn't need it
-    if (USTR(from) == u_common)
-        ok = osl_get_inverse_matrix (sg, &M, (const char *)to);
-    else if (USTR(to) == u_common)
-        ok = osl_get_matrix (sg, &M, (const char *)from);
-    else
-        ok = osl_get_from_to_matrix (sg, &M, (const char *)from,
-                                     (const char *)to);
-    if (ok) {
-        if (vectype == TypeDesc::POINT) {
-            if (Pin_derivs)
-                osl_transform_dvmdv(Pout, &M, Pin);
-            else
-                osl_transform_vmv(Pout, &M, Pin);
-        } else if (vectype == TypeDesc::VECTOR) {
-            if (Pin_derivs)
-                osl_transformv_dvmdv(Pout, &M, Pin);
-            else
-                osl_transformv_vmv(Pout, &M, Pin);
-        } else if (vectype == TypeDesc::NORMAL) {
-            if (Pin_derivs)
-                osl_transformn_dvmdv(Pout, &M, Pin);
-            else
-                osl_transformn_vmv(Pout, &M, Pin);
-        }
-        else ASSERT(0);
-    } else {
-        *(Vec3 *)Pout = *(Vec3 *)Pin;
-        if (Pin_derivs) {
-            ((Vec3 *)Pout)[1] = ((Vec3 *)Pin)[1];
-            ((Vec3 *)Pout)[2] = ((Vec3 *)Pin)[2];
-        }
-    }
-    if (Pout_derivs && !Pin_derivs) {
-        ((Vec3 *)Pout)[1].setValue (0.0f, 0.0f, 0.0f);
-        ((Vec3 *)Pout)[2].setValue (0.0f, 0.0f, 0.0f);
-    }
-    return ok;
-}
-
-
-
-OSL_SHADEOP int
-osl_transform_triple_nonlinear (void *sg_, void *Pin, int Pin_derivs,
-                                void *Pout, int Pout_derivs,
-                                void *from, void *to,
-                                int vectype)
-{
-    ShaderGlobals *sg = (ShaderGlobals *)sg_;
-    RendererServices *rend = sg->renderer;
-    if (rend->transform_points (sg, USTR(from), USTR(to), sg->time,
-                                (const Vec3 *)Pin, (Vec3 *)Pout, 1,
-                                (TypeDesc::VECSEMANTICS)vectype)) {
-        // Renderer had a direct way to transform the points between the
-        // two spaces.
-        if (Pout_derivs) {
-            if (Pin_derivs) {
-                rend->transform_points (sg, USTR(from), USTR(to), sg->time,
-                                        (const Vec3 *)Pin+1,
-                                        (Vec3 *)Pout+1, 2, TypeDesc::VECTOR);
-            } else {
-                ((Vec3 *)Pout)[1].setValue (0.0f, 0.0f, 0.0f);
-                ((Vec3 *)Pout)[2].setValue (0.0f, 0.0f, 0.0f);
-            }
-        }
-        return true;
-    }
-
-    // Renderer couldn't or wouldn't transform directly
-    return osl_transform_triple (sg, Pin, Pin_derivs, Pout, Pout_derivs,
-                                 from, to, vectype);
-}
-
-
-
-OSL_SHADEOP void
-osl_transpose_mm (void *r, void *m)
-{
-    MAT(r) = MAT(m).transposed();
-}
-
-// Calculate the determinant of a 2x2 matrix.
-template <typename F>
-inline F det2x2(F a, F b, F c, F d)
-{
-    return a * d - b * c;
-}
-
-// calculate the determinant of a 3x3 matrix in the form:
-//     | a1,  b1,  c1 |
-//     | a2,  b2,  c2 |
-//     | a3,  b3,  c3 |
-template <typename F>
-inline F det3x3(F a1, F a2, F a3, F b1, F b2, F b3, F c1, F c2, F c3)
-{
-    return a1 * det2x2( b2, b3, c2, c3 )
-         - b1 * det2x2( a2, a3, c2, c3 )
-         + c1 * det2x2( a2, a3, b2, b3 );
-}
-
-// calculate the determinant of a 4x4 matrix.
-template <typename F>
-inline F det4x4(const Imath::Matrix44<F> &m)
-{
-    // assign to individual variable names to aid selecting correct elements
-    F a1 = m[0][0], b1 = m[0][1], c1 = m[0][2], d1 = m[0][3];
-    F a2 = m[1][0], b2 = m[1][1], c2 = m[1][2], d2 = m[1][3];
-    F a3 = m[2][0], b3 = m[2][1], c3 = m[2][2], d3 = m[2][3];
-    F a4 = m[3][0], b4 = m[3][1], c4 = m[3][2], d4 = m[3][3];
-    return a1 * det3x3( b2, b3, b4, c2, c3, c4, d2, d3, d4)
-         - b1 * det3x3( a2, a3, a4, c2, c3, c4, d2, d3, d4)
-         + c1 * det3x3( a2, a3, a4, b2, b3, b4, d2, d3, d4)
-         - d1 * det3x3( a2, a3, a4, b2, b3, b4, c2, c3, c4);
-}
-
-OSL_SHADEOP float
-osl_determinant_fm (void *m)
-{
-    return det4x4 (MAT(m));
 }
